:recycle: notice-api 수정
